+++ tests/testthat/test-paths.R
@@ -1,7 +1,5 @@
 context("dodgr_paths")
 
-test_all <- (identical (Sys.getenv ("MPADGE_LOCAL"), "true"))
-
 test_that("paths", {
     graph <- weight_streetnet (hampi)
     from <- graph$from_id [1:100]