+++ tests/testthat/test-sc.R
@@ -1,6 +1,7 @@
 context("SC")
 
-test_all <- (identical (Sys.getenv ("MPADGE_LOCAL"), "true"))
+test_all <- (identical (Sys.getenv ("MPADGE_LOCAL"), "true") |
+             identical (Sys.getenv ("GITHUB_WORKFLOW"), "test-coverage"))
 
 skip_on_cran ()
 skip_if (!test_all)
@@ -96,8 +97,7 @@ test_that ("traffic light nodes", {
 
                expect_identical (net_sc0$d, net_sc1$d)
                expect_identical (net_sc0$d_weighted, net_sc1$d_weighted)
-               if (!identical (Sys.getenv ("TRAVIS"), "true"))
-                   expect_identical (net_sc0$time, net_sc1$time)
+               expect_true (!identical (net_sc0$time, net_sc1$time))
                expect_identical (net_sc0$time_weighted, net_sc1$time_weighted)
 
                expect_silent (net_sc1 <- weight_streetnet (hsc, wt_profile = 1))